refactor: altera pacotes do domain e do service
